requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.1
alembic==1.14.1
anaconda-anon-usage==0.4.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
boltons==23.0.0
Brotli==1.0.9
cattrs==24.1.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.8
colorlog==6.9.0
comm==0.2.2
conda==24.11.0
conda-content-trust==0.2.0
conda-libmamba-solver==24.1.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.3.0
cryptography==42.0.5
cycler==0.12.1
datasets==3.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.8.0
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.0
executing==2.0.1
fastf1==3.4.0
fastjsonschema==2.20.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.4
imbalanced-learn==0.13.0
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.24.0
ipywidgets==8.1.5
isoduration==20.11.0
iterative-stratification==0.1.9
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
libmambapy==1.5.8
lightning-utilities==0.11.9
lion-pytorch==0.2.3
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MedPy==0.5.2
menuinst==2.0.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.0
optuna==4.3.0.dev0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==23.3.1
platformdirs==3.10.0
pluggy==1.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.42
propcache==0.2.1
protobuf==5.28.3
psutil==5.9.8
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.0
pycosat==0.6.6
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.7
ray==2.41.0
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.6
ruamel.yaml==0.17.21
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.20.0
serpapi==0.1.5
setproctitle==1.3.4
setuptools==68.2.2
SimpleITK==2.4.1
simsimd==6.2.1
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.37
stack-data==0.6.2
stringzilla==3.11.3
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.14
timple==0.1.8
tinycss2==1.3.0
torch==2.3.0
torchaudio==2.3.0
torchinfo==1.8.0
torchmetrics==1.6.1
torchvision==0.18.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
truststore==0.8.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.1.0
wandb==0.19.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.1.1
wheel==0.41.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.17.0
zstandard==0.19.0
